Rewrite server concurrency
We can shorten much of the code and achieve better compositionality by forking
a reader thread to do all the reading. This means our main logic is entirely
pure STM so it can be interrupted using `race` or my `raceSTM`. Indeed this is
very close to the original approach in eeeaac4.